+++ src/krotov/result.py
@@ -3,8 +3,8 @@ import pickle
 import time
 from textwrap import dedent
 
+from .conversions import _nested_list_shallow_copy, pulse_onto_tlist
 from .objectives import Objective, _ControlPlaceholder
-from .structural_conversions import _nested_list_shallow_copy, pulse_onto_tlist
 
 
 __all__ = ['Result']